requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py37_0
backcall=0.2.0=py_0
blas=1.0=mkl
brotlipy=0.7.0=py37h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.1.19=hecd8cb5_1
certifi=2020.12.5=py37hecd8cb5_0
cffi=1.14.4=py37h2125817_0
chardet=4.0.0=py37hecd8cb5_1003
cryptography=3.3.1=py37hbcfaee0_0
cycler=0.10.0=py37_0
decorator=4.4.2=py_0
ffmpeg=4.2.2=h97e5cf8_0
freetype=2.10.4=ha233b18_0
gettext=0.19.8.1=hb0f4f8b_2
gmp=6.1.2=hb37e062_1
gnutls=3.6.5=h91ad68e_1002
idna=2.10=pyhd3eb1b0_0
intel-openmp=2019.4=233
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
jedi=0.17.2=py37_0
joblib=0.17.0=py_0
jpeg=9b=he5867d9_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37_0
kiwisolver=1.2.0=py37h04f5b5a_0
lame=3.100=h1de35cc_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.2=h79c402e_3
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h181983e_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
ncurses=6.2=h0a44026_1
nettle=3.4.1=h3018a27_0
ninja=1.10.1=py37h879752b_0
numpy=1.19.1=py37h3b9f5b6_0
numpy-base=1.19.1=py37hcfb5961_0
olefile=0.46=py37_0
opencv-python=4.4.0.44=pypi_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1k=h9ed2024_0
opt_einsum=3.1.0=py_0
pandas=1.1.3=py37hb1e8313_0
parso=0.7.0=py_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=8.0.0=py37h1a82f1a_0
pip=20.2.4=py37_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.6.0=py37_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=py_0
pytorch=1.8.0=py3.7_0
pytz=2020.1=py_0
pyzmq=19.0.2=py37hb1e8313_1
readline=8.0=h1de35cc_0
requests=2.25.1=pyhd3eb1b0_0
scikit-learn=0.23.2=py37h959d312_0
scipy=1.5.2=py37h912ce22_0
setuptools=50.3.0=py37h0dc7051_1
six=1.15.0=py_0
sqlite=3.33.0=hffcf06c_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.8.0=py37
torchvision=0.9.0=py37_cpu
tornado=6.0.4=py37h1de35cc_1
tqdm=4.50.2=py_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.3=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
wheel=0.35.1=py_0
x264=1!157.20191217=h1de35cc_0
xlrd=1.2.0=py37_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.3=hb1e8313_3
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
